refactor table items
