+++ processor/src/db/models/ans_models/ans_utils.rs
@@ -5,7 +5,7 @@
 #![allow(clippy::extra_unused_lifetimes)]
 
 use crate::{
-    db::models::new_default_models::move_resources::MoveResource,
+    db::models::default_models::move_resources::MoveResource,
     utils::util::{
         bigdecimal_to_u64, deserialize_from_string, parse_timestamp_secs, standardize_address,
         truncate_str,
