+++ processor/src/db/models/default_models/mod.rs
@@ -1,9 +1,6 @@
 pub mod block_metadata_transactions;
-pub mod current_table_items;
 pub mod move_resources;
 pub mod parquet_move_modules;
-pub mod parquet_move_tables;
 pub mod parquet_transactions;
 pub mod parquet_write_set_changes;
 pub mod table_items;
-pub mod table_metadata;
