+++ processor/src/db/models/default_models/move_resources.rs
