+++ processor/src/db/models/mod.rs
@@ -1,9 +1,8 @@
 pub mod ans_models;
 pub mod coin_models;
+pub mod default_models;
 pub mod fungible_asset_models;
-pub mod new_default_models;
 pub mod object_models;
-pub mod old_default_models;
 pub mod stake_models;
 pub mod token_models;
 pub mod token_v2_models;
