--- processor/src/db/models/old_default_models/mod.rs
@@ -1,4 +1,0 @@
-// Copyright © Aptos Foundation
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod postgres_move_tables;
