+++ processor/src/db/models/stake_models/stake_utils.rs
@@ -2,9 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
-    db::models::{
-        new_default_models::move_resources::MoveResource, token_models::token_utils::Table,
-    },
+    db::models::{default_models::move_resources::MoveResource, token_models::token_utils::Table},
     utils::util::{deserialize_from_string, standardize_address},
 };
 use anyhow::{Context, Result};
